Adding DelphiBridge service that fetches the current location for an old objectId + schemaName in a museum.
