+++ service_thing_aggregate/test/resources/evolutions/test/1.sql
@@ -44,6 +44,7 @@ CREATE TABLE MUSIT_MAPPING.MUSITTHING (
   subNoAsNumber    INTEGER,
   mainobject_id    NUMBER(20),
   old_schema_name  VARCHAR(500),
+  lokal_pk         NUMBER(20),
   PRIMARY KEY (object_id)
 );
